Create an Autoregressive Integrated Moving Average (ARIMA) Process

Sets up the necessary backend for the ARIMA process.

ARIMA(ar = 1, i = 0, ma = 1, sigma2 = 1)

Arguments

ar

A vector or integer containing either the coefficients for \(\phi\)'s or the process number \(p\) for the Autoregressive (AR) term.

i

An integer containing the number of differences to be done.

ma

A vector or integer containing either the coefficients for \(\theta\)'s or the process number \(q\) for the Moving Average (MA) term.

sigma2

A double value for the standard deviation, \(\sigma\), of the ARIMA process.

Value

An S3 object with called ts.model with the following structure:

process.desc

\(AR*p\), \(MA*q\)

theta

\(\sigma\)

plength

Number of Parameters

obj.desc

y desc replicated x times

obj

Depth of Parameters e.g. list(c(length(ar),length(ma),1) )

starting

Guess Starting values? TRUE or FALSE (e.g. specified value)

Details

A variance is required since the model generation statements utilize randomization functions expecting a variance instead of a standard deviation like R.

Examples

# Create an ARMA(1,2) process
ARIMA(ar=1,2)
# Creates an ARMA(3,2) process with predefined coefficients.
ARIMA(ar=c(0.23,.43, .59), ma=c(0.4,.3))

# Creates an ARMA(3,2) process with predefined coefficients and standard deviation
ARIMA(ar=c(0.23,.43, .59), ma=c(0.4,.3), sigma2 = 1.5)