Find the auto imu result

Provides the core material to create an S3 object for auto.imu

auto_imu_cpp(data, combs, full_model, alpha, compute_v, model_type, K, H, G,
  robust, eff, bs_optimism, seed)

Arguments

data

A mat containing multiple columns of independent data with the same number of observations.

full_model

A vector<string> that contains the largest / full model.

alpha

A double that indicates the alpha level for CIs.

compute_v

A string indicating the type of V matrix to generate

model_type

A string that describes the model generation / transformation: 'ssm' or 'imu'

K

A int that controls how many times the GMWM is run.

H

A int that controls how many bootstraps occur.

G

A int that controls how many guesses occur.

robust

A bool that indicates whether to use classical or robust wavelet variance.

eff

A double that indicates the efficiency to use.

bs_optimism

A bool that indicates whether the model selection score should be calculated with bootstrap or asymptotics.

seed

A unsigned int that is the seed one wishes to use.

model_str

A vector<vector<string>> that gives a list of models to test.

Value

A field<field<field<mat>>> that contains the model score matrix and the best GMWM model object.