Drift to WV

This function compute the WV (haar) of a Drift process

dr_to_wv(omega, tau)

Arguments

omega

A double corresponding to the slope of the drift

tau

A vec containing the scales e.g. \(2^{\tau}\)

Value

A vec containing the wavelet variance of the drift.

Process Haar Wavelet Variance Formula

The Drift (DR) process has a Haar Wavelet Variance given by: $$ \nu _j^2\left( {{\omega }} \right) = \frac{{\tau _j^2{\omega ^2}}}{{16}} $$

Haar Wavelet Derivation Information

For more information, please see: http://smac-group.com/computing/process-to-haar-wavelet-variance-formulae (Internet Connection Required).

Examples

ntau = 8
tau = 2^(1:ntau)
wv.theo = dr_to_wv(-2.3, tau)