Generate Autoregressive Order

Generate an ARMA(\(p\),\(q\)) process with supplied vector of Autoregressive Coefficients (\(\phi\)), Moving Average Coefficients (\(\theta\)), and \(\sigma^2\).

gen_arma(N, ar, ma, sigma2 = 1.5, n_start = 0L)

Arguments

N

An integer for signal length.

ar

A vec that contains the AR coefficients.

ma

A vec that contains the MA coefficients.

sigma2

A double that contains process variance.

n_start

An unsigned int that indicates the amount of observations to be used for the burn in period.

Value

A vec that contains the generated observations.

Details

For AR(1), MA(1), and ARMA(1,1) please use their functions if speed is important as this function is designed to generate generic ARMA processes.

Process Definition

The Autoregressive order \(p\) and Moving Average order \(q\) (ARMA(\(p\),\(q\))) process with non-zero parameters \(\phi_i \in (-1,+1)\) for the AR components, \(\theta_j \in (-1,+1)\) for the MA components, and \(\sigma^2 \in {\rm I\!R}^{+}\). This process is defined as:

$${X_t} = \sum\limits_{i = 1}^p {{\phi _i}{X_{t - i}}} + \sum\limits_{i = 1}^q {{\theta _i}{\varepsilon _{t - i}}} + {\varepsilon _t}$$ where $${\varepsilon_t}\mathop \sim \limits^{iid} N\left( {0,\sigma^2} \right)$$

Generation Algorithm

The innovations are generated from a normal distribution. The \(\sigma^2\) parameter is indeed a variance parameter. This differs from R's use of the standard deviation, \(\sigma\).

Examples

gen_arma(10, c(.3,.5), c(.1), 1, 0)

#>            [,1]
#>  [1,] 2.0567729
#>  [2,] 1.9164544
#>  [3,] 1.0638526
#>  [4,] 1.3836825
#>  [5,] 1.2956966
#>  [6,] 1.3460503
#>  [7,] 0.8289425
#>  [8,] 0.8546455
#>  [9,] 0.4448576
#> [10,] 1.7887429