Retrieve GMWM starting value from Yannick's objective function

Obtains the GMWM starting value given by Yannick's objective function optimization

getObjFun(theta, desc, objdesc, model_type, omega, wv_empir, tau)

Arguments

theta

A vec with dimensions N x 1 that contains user-supplied initial values for parameters

desc

A vector<string> indicating the models that should be considered.

objdesc

A field<vec> containing a list of parameters (e.g. AR(1) = c(1,1), ARMA(p,q) = c(p,q,1))

model_type

A string containing the model type. Either 'imu' or 'ssm'

omega

A mat that is the inverse of the diagonal of the V matrix.

wv_empir

A vec containing the empirical wavelet variance.

tau

A vec that contains the scales of 2^(1:J), where J is the number of scales created by the decomposition.

Value

A double that is the value of the Objective function under Yannick's starting algorithm