Routing function for summary info

Gets all the data for the summary.gmwm function.

get_summary(theta, desc, objdesc, model_type, wv_empir, theo, scales, V, omega,
  obj_value, N, alpha, robust, eff, inference, fullV, bs_gof, bs_gof_p_ci,
  bs_theta_est, bs_ci, B)

Arguments

theta

A vec with dimensions N x 1 that contains user-supplied initial values for parameters

desc

A vector<string> indicating the models that should be considered.

objdesc

A field<vec> containing a list of parameters (e.g. AR(1) = c(1,1), ARMA(p,q) = c(p,q,1))

model_type

A string that represents the model transformation

wv_empir

A vec that

theo

A vec that

scales

A vec that

V

A mat that contains the V matrix used to obtain the GMWM.

omega

A mat that

obj_value

A double that contains the objective function value at the optimized solution.

N

A int that indicates how long the time series is.

alpha

A double that handles the alpha level of the confidence interval (1-alpha)*100

robust

A bool that indicates whether the estimation should be robust or not.

eff

A double that specifies the amount of efficiency required by the robust estimator.

inference

A bool that indicates whether inference (e.g. GoF) should be run.

fullV

A bool that indicates whether the matrix has been fully bootstrapped.

bs_gof

A bool indicating whether the GoF should be bootstrapped or done asymptotically.

bs_gof_p_ci

A bool indicating whether a bootstrapped p-value should be generated during the bootstrapped GoF

bs_ci

A bool that indicates whether a bootstrapped CI should be obtained or to use analytical derivatives.

B

A int that indicates how many iterations should take place.

Value

A field<mat> that contains bootstrapped / asymptotic GoF results as well as CIs.