Predict future points in the time series using the solution of the

Creates a prediction using the estimated values of GMWM through the ARIMA function within R.

# S3 method for gmwm
predict(object, data.in.gmwm, n.ahead = 1, ...)

Arguments

object

A gmwm object

data.in.gmwm

The data SAME EXACT DATA used in the GMWM estimation

n.ahead

Number of observations to forecast

...

Additional parameters passed to ARIMA Predict

Value

A predict.gmwm object with:

pred

Predictions

se

Standard Errors

resid

Residuals from ARIMA ML Fit

See also

gmwm, ARMA

Examples

# Simulate an ARMA Process
xt = gen_gts(1000, ARMA(ar=0.3, ma=0.6, sigma2=1))
model = gmwm(ARMA(1,1), xt)

# Make prediction
predict(model, xt, n.ahead = 1)

#> $pred
#> Time Series:
#> Start = 1001 
#> End = 1001 
#> Frequency = 1 
#> [1] 1.217753
#> 
#> $se
#> Time Series:
#> Start = 1001 
#> End = 1001 
#> Frequency = 1 
#> [1] 0.9851801
#> 
#> $resid
#> Time Series:
#> Start = 1 
#> End = 1000 
#> Frequency = 1 
#>    [1] -1.2432490407  0.0765037213 -0.4948146692 -0.8380659570  1.7428749271
#>    [6] -2.6608470670  0.3210624758  1.4609740043 -1.5669864276 -0.6708744718
#>   [11] -0.2349524726 -1.5017710834 -0.8610613761 -0.7667105899 -0.1729424780
#>   [16] -1.0333267985  0.9741995938  1.1485571330 -1.3410734850  0.4610016313
#>   [21] -0.4397962057  0.0986688349 -2.0997471782 -1.7344105549  1.3426843976
#>   [26]  0.1949020915  0.2275471659  0.1929272401  1.0490848861  0.2003014613
#>   [31] -0.4886289423 -0.1569033110  0.3123255694 -0.4541274953 -0.7698680347
#>   [36]  0.6363959153 -1.3065655943 -0.8740470287 -0.8340245435  1.1155164663
#>   [41] -0.0340983750  1.3811849180 -0.0976029395  0.8367451019  0.5136134891
#>   [46] -0.6410756518  1.0972970536  0.0860021121 -0.8640904913  1.0318458979
#>   [51]  0.9926802999  1.3529563234  0.0492642169  0.7075212349  0.8801698332
#>   [56] -0.1521635397 -0.1899227010  1.4839709465  1.3842138158 -0.6598606635
#>   [61]  0.0366225114  0.0472105585 -0.9743653950 -0.4691452614 -1.1361258102
#>   [66]  1.0713591013  1.4515737264  0.8013426514 -0.7146496742 -0.1438270375
#>   [71] -0.5897907653 -0.2925717623  0.2526877392  0.1994273847  0.0112916952
#>   [76] -0.1707498826 -1.3304195022  0.4560239963 -1.1614983573  0.1970116836
#>   [81] -0.6449687305  0.4616110754  0.8439131590  0.6391107319 -0.5096200670
#>   [86]  0.4171335905 -1.8294129425 -0.0534171159  0.5191510299 -0.2451570616
#>   [91] -0.4263562683  0.8721509194 -0.1884271652 -0.1836029332 -0.9096068396
#>   [96]  0.5047465367  0.4121615872 -1.0231818937 -0.8025975915  0.1111882559
#>  [101] -1.4604008313  1.5528862645  0.4231862936  0.3996466171  0.9160698389
#>  [106]  0.7150619048  0.0891206638 -0.7170626131  0.4018832351 -0.7136949667
#>  [111]  1.5688596728 -0.2235730204  0.2308170367 -1.8626064153 -1.9197668975
#>  [116]  0.7325285885 -1.1476624269  0.2888344909 -0.6632617380  0.4147486802
#>  [121]  0.7649005971 -1.1714276329 -0.8464108279 -0.3824299760  1.0901175847
#>  [126] -2.2084986914  1.5081289561  1.5084518861 -1.1072394193  1.1813683469
#>  [131]  0.8278796534 -0.0129381359  0.6608163561  0.4291095927 -1.5426276035
#>  [136]  0.7771331182 -1.1199614095  0.1303675563 -1.4907897499 -1.9389679004
#>  [141] -0.4843956105  0.9617412118  1.7297289608  2.1681926180  0.7646986581
#>  [146] -1.1161567601  0.2892441825  0.9446826461 -0.8441401197 -2.5940782131
#>  [151]  0.3205079987  0.4851319691 -1.8135307894 -1.7238113920  0.1036658761
#>  [156] -1.9402863329  0.8492300983 -0.8656952246 -0.6922351196 -0.8251778045
#>  [161] -0.9954902069 -0.1956150503  1.2593393794 -0.0839485555 -0.1033727202
#>  [166] -1.2229166801 -0.1226057315 -1.2524551909  0.5177362603 -0.4718423197
#>  [171]  0.8246791460 -0.7424784523 -0.6880061994 -0.9269024595  0.8554431345
#>  [176]  1.2952112663  0.0620263022  1.2426899709  1.4844247822  0.5474401680
#>  [181] -0.8375698253  0.7450067874 -0.6851599959 -0.7517010282 -0.4094235272
#>  [186]  1.1891324866 -0.3865575414 -2.2494900798 -0.8135386785 -0.3003968006
#>  [191] -1.7177016446  0.3881085334  0.1452200784 -0.2791832202  1.4049806280
#>  [196]  0.4798299069 -0.8175165298  0.0334251164  0.1215362756 -0.1560883009
#>  [201]  1.2282556639 -0.2254843594  1.0536700174  1.1020560447 -0.0012526976
#>  [206] -0.7813206916  0.6309648207  1.4765780048  0.5061849135  0.4190857665
#>  [211] -0.8913825138 -1.7881015930 -0.7208328069  0.8737638660  0.4593527557
#>  [216]  0.7147717631 -0.8106420885 -2.0341087103  0.2015534904  1.0559027169
#>  [221] -0.7773763787  0.3947542927 -1.4940422596 -0.2139878809  1.2708962783
#>  [226] -0.5314145835  0.3639860343 -0.0323453123  0.2891019475 -0.8150912773
#>  [231] -0.4593857261 -1.4526078307 -0.6200273914  1.8244215583 -1.7034439346
#>  [236]  1.0828068975  1.7440924437 -0.4490929707 -1.8011852003  0.1377211969
#>  [241]  1.1762998223  0.4085830881  0.4311819433  1.2028161320  0.6624194005
#>  [246] -1.2518236171 -0.1560926510 -1.1486632461 -0.4840782546 -0.8261422020
#>  [251]  0.4269377613  0.0272184007  0.5346007534 -0.1938126731  0.5565260194
#>  [256] -0.0516532545  0.9135584453 -1.9692124431  0.0135600644  1.4904481002
#>  [261]  0.9934055201  0.1065038898  0.6638412244 -2.1011325899  0.9754104681
#>  [266] -0.7903879819  0.4421154797  1.1765413954 -0.4247789076 -0.1230122059
#>  [271] -0.2984952365  0.0944806197 -0.1669313105 -0.1800005949  1.7575409551
#>  [276]  0.7341158908 -1.6776900034 -0.0632123430  0.6338009938  1.7516181409
#>  [281]  0.7424587552 -0.1110060848  1.4581752385 -0.0823955117  1.6942465203
#>  [286] -0.0313755307  1.5474749507  0.6091389360 -0.3830787298 -0.9705484114
#>  [291] -0.6450419926 -0.4185953817 -0.7801418475 -0.9714944862 -1.2973347959
#>  [296] -1.9538489298 -0.9922844500 -0.3272620221 -0.2759598447 -0.6788445194
#>  [301]  0.4930644788  0.5863530620  0.1380124345  0.5060637549 -0.4726312078
#>  [306] -0.1300089138 -0.7724971487 -0.6216741356  1.7094013769  0.6220222889
#>  [311] -0.3642340847  0.9881949233  1.0084437824  1.4426705032 -0.8756865147
#>  [316]  1.4015464860  0.9579246883 -0.6088219028  0.5392395399  0.8919783259
#>  [321]  0.4164386638 -0.4034435002  0.4910689200 -1.5636822401 -1.4435098326
#>  [326] -0.0850354035 -0.9364914604 -0.2500946373  0.5277105442 -1.2151376398
#>  [331]  0.0329158193  0.7433017123 -0.4263478927 -0.4924285032 -0.7407234983
#>  [336]  3.0008104375 -0.4734979421 -1.4452423605 -0.4633981520 -0.0288770838
#>  [341] -0.0072882265  0.5546179291 -0.2665457981 -1.0281385877 -0.8296508210
#>  [346]  0.1574420085  0.0005920936 -2.5180712074  0.5041599801  1.6354866240
#>  [351]  0.9109339556  1.5335432209 -1.9140059441 -1.2847017256  2.1315519805
#>  [356] -0.7502644741  0.8910016161 -0.7364361090 -0.7171956216 -0.5383475328
#>  [361] -0.1049725763 -1.3594163631 -1.2055945072 -1.2060294587 -0.9917106900
#>  [366]  1.4358022252 -1.0738462680  0.6558145558  1.1094393434 -0.3393032053
#>  [371] -0.6606574559  0.1206024185 -0.0563620410  1.6159305486  0.6593621719
#>  [376]  0.1628915227 -0.4482448557  0.6305097770 -0.2149690068 -0.0019872540
#>  [381] -1.8987280717 -0.4934005682  0.4971271135  0.4532217639 -0.2870134123
#>  [386] -0.3637303783 -0.8824650785 -1.0255107341  0.8872297601  0.0698794101
#>  [391]  0.8101952720  0.3342410989 -1.1313738065 -0.5081291331  3.1030544327
#>  [396] -0.5735064557  0.1597811395  0.0770127086  0.2153346930  1.0489491095
#>  [401] -0.3785663702  0.1794083858  0.6108035053 -0.2519693796 -0.0204556053
#>  [406] -0.1862770116  0.0100293199  1.4503972900  0.1409727474  0.4217363790
#>  [411]  0.8912827724  0.3506244078  0.8683231895 -0.9596781962  0.0538785212
#>  [416] -0.6041974183  1.4871652827 -0.0103677750  0.0262021884 -0.1957840883
#>  [421] -2.5162740888  1.4920806860  1.5395892730  0.6447528728 -1.3759114864
#>  [426]  2.5986748739 -0.1148820484  0.6632764781  0.3916325303  0.5426770668
#>  [431] -0.2786960360  0.0581514883  0.5953940986 -0.9416457923 -0.7413872531
#>  [436]  0.3518562505 -1.7567039494 -0.4761381147 -1.1386716357 -0.4629604685
#>  [441] -0.5242958586 -1.3761517931 -1.5319084251 -0.4975105278  1.5019738521
#>  [446]  0.3222791838  0.5500142914  0.8575144612  0.6021316628 -1.6076648717
#>  [451] -0.2166747136  0.3627444149  1.5794711556  0.4218107476  2.0402578300
#>  [456] -0.3630182964 -1.7106157158 -0.9372965117  0.7575559930 -0.0277590739
#>  [461]  0.7049114168  0.1730323251  0.1255136022  0.8061136757 -0.5691154189
#>  [466]  0.1039435481  0.7026573495  0.7407189322 -1.6690682344 -0.1161203085
#>  [471] -0.1476048974 -0.7043207264 -0.2767106274  0.4100789245  1.0661092969
#>  [476]  0.1652366114  0.7066135120  0.2194112955 -1.1807316910 -1.3572804675
#>  [481]  0.6441097578  0.1571352553  0.6710968924  0.0083565958 -0.0170058627
#>  [486]  0.3872127596  1.0729782311 -0.2835353847 -0.3182856356  0.1711231799
#>  [491] -1.4496086203  0.6770334441 -0.2293830512 -0.8849288276 -1.4114864131
#>  [496] -0.2418403883  1.6292688379  0.5915489178 -0.3894218343 -1.0278017656
#>  [501]  0.7608051909 -1.6048565306  0.2786254156 -1.0297096247 -0.5576408918
#>  [506]  0.0602703350 -0.4942640814 -0.1281059840  0.6212343971  0.3346689457
#>  [511]  0.2355899372  1.2432300877 -1.3577366538 -1.1592211912  0.9018257576
#>  [516]  1.7666428473  0.1665888381 -0.9769926569  0.9900698651 -1.6801115719
#>  [521]  2.4701733836 -0.2597703570  0.0037935204  0.4949096378 -0.3129151898
#>  [526] -2.5091688648  0.0454082362 -0.9714970867 -1.1539196849  1.1074861288
#>  [531]  1.0329502201 -0.8670148354  0.4729871668  0.0281637039  0.6674755931
#>  [536]  0.6431207708 -0.7249167411  0.1165658829 -0.3553564326 -0.4147284157
#>  [541] -0.4672180076  1.2903627123 -0.8423991858 -0.3159976852  0.5821299939
#>  [546] -0.3704216256  0.3597675077 -0.3637012320 -1.3156574879  0.4093265335
#>  [551] -1.1269195049 -0.3770229000 -0.1191448520 -0.2104203596  0.1456540117
#>  [556]  0.0453264163  0.5889492652  0.6571059503  0.5320073634 -0.4369827136
#>  [561] -0.2725824135  1.2495530375 -0.5103728070  1.5411637556  0.6125476626
#>  [566] -1.1049431548  1.8444645180 -1.1023343090  0.3665062314 -0.5582516362
#>  [571] -1.0726240451  0.4177162347 -1.5185088029 -0.6923701921 -0.2426534302
#>  [576] -2.6673043883 -1.1345332241  0.4498599191  2.1895866394  0.3171234904
#>  [581]  0.7038291871 -0.0172014276 -0.1121610457  0.3721008928  0.3843866100
#>  [586] -0.0708917589  0.6387911563  0.5661971286  1.0316393588 -0.5790844998
#>  [591] -1.4163156891  0.8410737219 -0.8865596058 -0.6389013677 -0.7782575461
#>  [596]  0.1226622678  2.6661215642 -1.2671711570  1.6464423656  1.8695666845
#>  [601]  0.8074879445 -0.2661274391 -1.1214703570 -1.4008178346  0.1395313065
#>  [606] -1.6280893548 -0.8753561480 -0.2960231510 -0.3292379024 -0.6404375857
#>  [611]  1.2227211381  1.0950913047  1.0855618538 -0.1884872599 -1.3586887484
#>  [616] -1.0364954663  0.0697145083  0.1605895372  0.9188566040 -0.0200904685
#>  [621] -1.7173547134 -1.5937638389  1.8068814082 -0.0861298197  0.2832591332
#>  [626] -0.5979595614 -0.2041512397  0.1034756454  0.0909360368  0.1730257263
#>  [631] -0.9439230646  0.1624701908 -0.6819144431  0.2722343800  1.2219030582
#>  [636] -0.3829124979 -1.6144355306 -1.0248767717 -1.4790360527  1.2334109383
#>  [641] -1.7682349736 -1.7325096735 -0.9046247970  0.0018285536 -1.7415102416
#>  [646] -1.0897452146  0.3648309244  0.9393299412 -0.7294232562  0.5882291501
#>  [651] -1.5134216861 -0.2403662401 -0.4258636323 -0.1848678885  1.0487902949
#>  [656] -0.7182832314 -0.5276812172 -1.0982411645  0.3877628900 -0.3422620125
#>  [661] -0.0444191678  0.1327322902  0.5804782100  0.9340463777  0.8091604934
#>  [666]  0.1917978125 -0.2487532808  1.1265704422  0.4280226117  0.6695514746
#>  [671] -0.9902398917  1.0196373898  1.0060217637  1.3470691142 -0.2421657874
#>  [676]  0.3937253413 -1.2906756482 -0.0941085254 -0.9993810087 -0.7469293300
#>  [681]  0.9955044339  1.2086975555  1.1113008110  0.6494717170 -0.8445715771
#>  [686] -1.3375169881 -0.4671277136 -0.4882981128  0.5901499253  1.5818925402
#>  [691] -0.2391810865 -0.4156325996  2.0996178933 -0.4038605841 -1.3301621251
#>  [696]  0.2591760593 -2.1123170422  0.3823378580 -0.1669855207 -1.7102808701
#>  [701] -0.2695810886 -1.9188256013  0.3757282619 -1.1212756985 -0.0375394148
#>  [706] -0.3193070138 -0.4923028042  0.7868751534 -0.5279710574 -0.9368397714
#>  [711] -1.5899329945 -0.5767736394  0.1040615129  1.2127465973  0.8358393523
#>  [716] -0.4880821184  0.8966170364 -0.1519788515 -0.3751750065  0.7752964827
#>  [721]  0.6222233468 -0.8730557184  0.7618053841  0.5230432388 -0.3032267558
#>  [726] -2.0404811734 -0.6052229550 -0.9465756041 -1.3637328477 -0.5471686812
#>  [731]  0.6782243576  0.5742986934  0.4896391951  0.1095841937 -0.1712086066
#>  [736]  0.5170860916 -1.6891677970  0.5856197694  0.2620130166 -0.8205085240
#>  [741] -0.4296838046 -0.4939983237  0.1401766178  0.6979681612 -0.9388091154
#>  [746]  0.1431232932  0.5995936333  0.5384856190  0.0766684788  1.2280113730
#>  [751]  0.6219862729  1.9180380991  0.3301021793  1.1309049085 -1.5376713266
#>  [756]  2.8893851087  0.5197448924  0.8552641068 -1.0906204944  0.5986964000
#>  [761] -0.5204795387 -0.7346952950  1.1765311632 -0.1903969042  2.0616165147
#>  [766] -0.3328513507  1.8366523508 -1.1998131116 -0.0983113351 -1.2021908439
#>  [771] -0.4662387220 -2.5422795350 -1.8378330108  0.1888457793 -0.0138172900
#>  [776] -0.4912406484  1.6275185028  1.0045329664 -0.7783053341  0.4636192820
#>  [781]  0.9540880894  1.0265170256  1.3051812745 -0.4244843866 -0.3059068553
#>  [786] -0.7391936990  1.8583953314  0.1335359181 -0.0431083529 -0.0425384270
#>  [791]  0.1717760990  1.1840467755  0.0120661154  0.8006266171 -0.3814589622
#>  [796] -0.7834442983  1.1771901905  1.4776174894  0.1811575421 -0.2494705601
#>  [801]  0.9129265401 -0.7281882129 -1.1813414811 -0.7656214679 -1.0157509608
#>  [806]  0.1478273094 -0.3421652240  2.2086946163  0.9520685609 -1.1894714383
#>  [811]  1.5402878602  0.4705647182 -1.5291959864  1.7332588570  0.3891830373
#>  [816]  0.1169992925  0.9938691904 -1.3170357947  0.1796564601 -1.3916472448
#>  [821] -0.8096154027  1.3711232180  0.1083229383 -3.1101115966  0.4032316767
#>  [826]  2.4171707446  1.7730197867 -0.1432517401 -0.1624284999 -0.6432497024
#>  [831] -0.3182930653  0.4043328102  0.8011654269 -1.4300018531 -0.7212625929
#>  [836]  2.3612657458 -1.0127775403 -1.0812454853 -0.2924878642  1.4325684589
#>  [841]  1.1558502541 -0.4225793599  2.3700276134  1.0122452190  1.2256526281
#>  [846]  1.2639051149 -0.0469061963  0.7373189364 -0.3633682362  1.0383641958
#>  [851]  0.5896031621 -0.5241216072 -0.1055738927 -0.4124256776  0.0567199412
#>  [856]  0.9873113575 -1.5593451553 -2.1640105874  0.2221491363  1.2406720326
#>  [861] -0.5774956828  0.2336184402  0.2046922000  0.0573067849 -0.2168600245
#>  [866] -0.0025380964  1.1478050754  0.1798225097  0.2517029746 -0.1840452312
#>  [871] -1.8860940802  1.6135993171 -0.3442803241 -0.8067144298 -0.9944526328
#>  [876] -0.4082899862 -0.4133622060 -1.6971381401 -0.0456640288  0.7635749574
#>  [881] -0.6127232033 -1.1404363180  0.4509862231  0.2995677349  0.6072331923
#>  [886] -0.6040896997  0.6736650627 -1.3748100558 -0.5384327253  0.1214857738
#>  [891] -0.1717299570 -0.8811727279  0.0691511566  1.2910764088  0.7195411851
#>  [896]  2.8675315071 -1.5910532292 -0.0299698366  1.7695262763 -1.8334586984
#>  [901] -0.0521293309  2.2233862479 -1.2412346655  0.9112654012  0.9230046471
#>  [906] -1.9341338741  1.1351843102  1.3543653179 -0.1350075795 -1.1380067285
#>  [911]  0.9353970667  1.3360131275  1.4529443098  0.6854110871 -1.4225843151
#>  [916] -1.0384289413  0.4731062129  1.8757243532 -0.5839450386 -1.5085777632
#>  [921]  0.2465056684  0.4396483862 -1.0658845893 -1.4611974773 -0.5099334754
#>  [926]  1.8268542488 -1.4917971447  0.1962276993 -0.6861084932  0.4109669705
#>  [931] -0.4904563996  1.6607576140  1.2056911307 -0.4150983324  0.4686884008
#>  [936]  0.8096225653 -0.4783617926 -0.3827555119  0.5864000891 -0.1161883796
#>  [941] -0.3095272812  0.6272458463 -0.5769912876  0.2622718775 -0.7195518679
#>  [946]  1.4062798293 -0.6908821808 -0.4151320900  0.2811681418 -1.1272216591
#>  [951]  0.8499099662 -1.0567126391  0.9723625056  0.2928746102  1.6212717183
#>  [956] -0.0366559367  0.2512760373  1.9218323766 -0.3353794294  1.7555609618
#>  [961]  0.0565308372  0.5692257097  0.5742504360 -0.2512462639  0.2492207661
#>  [966]  0.6405050561  0.5215598454  1.2467900116  0.6798008420  0.6723614548
#>  [971]  0.6953652509  1.0767017420 -0.6703619050 -0.1807567526 -1.4207865079
#>  [976]  0.9231537286  0.4798993128  0.7325146489 -0.5046324879 -0.2317104741
#>  [981] -0.3773528793 -0.4679079359 -0.8451383571  0.4870945246  0.6884864660
#>  [986] -1.2384143591 -0.4735434035 -0.8691407630  0.2466040633  2.0734213371
#>  [991]  0.2037120944 -0.7159751364  0.5799397138  0.3946140691 -2.3327648659
#>  [996]  1.3712566631 -0.6118782829 -2.2028023876 -0.5079309442  1.7885431865
#> 
#> attr(,"class")
#> [1] "predict.gmwm"