Wavelet Variance

Calculates the (MODWT) wavelet variance

wvar(x, ...)

# S3 method for lts
wvar(x, decomp = "modwt", filter = "haar", nlevels = NULL,
  alpha = 0.05, robust = FALSE, eff = 0.6, to.unit = NULL, ...)

# S3 method for gts
wvar(x, decomp = "modwt", filter = "haar", nlevels = NULL,
  alpha = 0.05, robust = FALSE, eff = 0.6, to.unit = NULL, ...)

# S3 method for ts
wvar(x, decomp = "modwt", filter = "haar", nlevels = NULL,
  alpha = 0.05, robust = FALSE, eff = 0.6, to.unit = NULL, ...)

# S3 method for default
wvar(x, decomp = "modwt", filter = "haar",
  nlevels = NULL, alpha = 0.05, robust = FALSE, eff = 0.6, freq = 1,
  from.unit = NULL, to.unit = NULL, ...)

# S3 method for imu
wvar(x, decomp = "modwt", filter = "haar", nlevels = NULL,
  alpha = 0.05, robust = F, eff = 0.6, to.unit = NULL, ...)

Arguments

x

A vector with dimensions N x 1, or a lts object, or a gts object, or a imu object.

...

Further arguments passed to or from other methods.

decomp

A string that indicates whether to use the "dwt" or "modwt" decomposition.

filter

A string that specifies what wavelet filter to use.

nlevels

An integer that indicates the level of decomposition. It must be less than or equal to floor(log2(length(x))).

alpha

A double that indicates the \(\left(1-p\right)*\alpha\) confidence level

robust

A boolean that triggers the use of the robust estimate.

eff

A double that indicates the efficiency as it relates to an MLE.

to.unit

A string indicating the unit which the data is converted to.

freq

A numeric that provides the rate of samples.

from.unit

A string indicating the unit which the data is converted from.

Value

A list with the structure:

"variance"

Wavelet Variance

"ci_low"

Lower CI

"ci_high"

Upper CI

"robust"

Robust active

"eff"

Efficiency level for Robust

"alpha"

p value used for CI

"unit"

String representation of the unit

Details

If `nlevels` is not specified, it is set to floor(log2(length(x)))

Examples

set.seed(999)
x = rnorm(100)
# Default
wvar(x)

#>       Variance      Low CI   High CI
#> 2  0.452177820 0.316045047 0.7004281
#> 4  0.230339706 0.140751059 0.4440086
#> 8  0.137102522 0.069880858 0.3811887
#> 16 0.051603837 0.020543745 0.2878292
#> 32 0.011723825 0.003282007 0.3685303
#> 64 0.008205545 0.001633306 8.3553640

# Robust
wvar(x, robust = TRUE, eff=0.3)

#>       Variance      Low CI    High CI
#> 2  0.549798331 0.247597104  1.1008892
#> 4  0.239938797 0.069556677  0.6463001
#> 8  0.131799310 0.013817156  0.5602001
#> 16 0.061956782 0.013817156  0.5797696
#> 32 0.007484506 0.003742253  0.4233625
#> 64 0.009967258 0.003742253 18.5216674

# 90% confidence interval
wvar(x, alpha = 0.10)

#>       Variance      Low CI   High CI
#> 2  0.452177820 0.334461347 0.6516656
#> 4  0.230339706 0.152096918 0.3978812
#> 8  0.137102522 0.077666345 0.3200430
#> 16 0.051603837 0.023732362 0.2119612
#> 32 0.011723825 0.004018555 0.1908247
#> 64 0.008205545 0.002136049 2.0867886

# IMU Object
## Not run: ------------------------------------
# if(!require("imudata")){
#    install_imudata()
#    library("imudata")
# }
# 
# data(imu6)
# test = imu(imu6, gyros = 1:3, accels = 4:6, freq = 100)
# df = wvar.imu(test)
## ---------------------------------------------